Cleanup dev module loading
