+++ packages/chookscord/src/scripts/dev/modules/commands.ts
@@ -1,6 +1,7 @@
 import * as lib from '@chookscord/lib';
 import type { ChooksCommand, ChooksSlashCommand } from '@chookscord/types';
 import type { Consola } from 'consola';
+import type { Module } from '../load-modules';
 import { basename } from 'path';
 import { createTimer } from '../../../utils';
 
@@ -37,7 +38,7 @@ export async function update(
   }
 }
 
-export function unlink(
+export function remove(
   paths: Record<string, string>,
   store: Store,
   filePath: string,
@@ -48,4 +49,19 @@ export function unlink(
 
   store.delete(commandName);
   logger?.success(`Command "${commandName}" deleted.`);
+}
+
+export function createModule(
+  store: Store,
+  logger?: Consola,
+): Module {
+  const paths = {};
+  return {
+    async compile(filePath: string) {
+      await update(paths, store, filePath, logger);
+    },
+    unlink(filePath: string) {
+      remove(paths, store, filePath, logger);
+    },
+  };
 }
