+++ packages/chookscord/src/scripts/dev/modules/subcommands.ts
@@ -2,6 +2,7 @@ import * as lib from '@chookscord/lib';
 import * as utils from '../../../utils';
 import type { ChooksCommand, ChooksSubCommand } from '@chookscord/types';
 import type { Consola } from 'consola';
+import type { Module } from '../load-modules';
 import { basename } from 'path';
 
 type Store = lib.Store<ChooksCommand>;
@@ -35,7 +36,7 @@ export async function update(
   }
 }
 
-export function unlink(
+export function remove(
   paths: Record<string, string>,
   store: Store,
   filePath: string,
@@ -46,4 +47,19 @@ export function unlink(
 
   store.delete(commandName);
   logger?.success(`Command "${commandName}" deleted.`);
+}
+
+export function createModule(
+  store: Store,
+  logger?: Consola,
+): Module {
+  const paths = {};
+  return {
+    async compile(filePath: string) {
+      await update(paths, store, filePath, logger);
+    },
+    unlink(filePath: string) {
+      remove(paths, store, filePath, logger);
+    },
+  };
 }
