+++ packages/chookscord/src/tools/index.ts
@@ -1,5 +1,6 @@
 export * from './changed';
 export * from './client';
+export * from './command';
 export * from './compile';
 export * from './config';
 export * from './find-files';
